+++ testcontainers/src/core/env/config.rs
@@ -17,6 +17,9 @@ pub enum ConfigurationError {
     #[cfg(feature = "properties-config")]
     #[error("failed to load testcontainers properties: {0}")]
     WrongPropertiesFormat(#[from] serde_java_properties::de::Error),
+    #[cfg(feature = "reusable-containers")]
+    #[error("container name or labels must be provided when reusing containers")]
+    MissingContainerNameAndLabels,
 }
 
 /// The default path to the Docker configuration file.